Combine servers into 1 file.
