+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "test": "npm run test-manifest",
     "test-manifest": "cross-env NODE_ENV=test mocha -r esm --preserve-symlinks -t 30000 -A -R ${REPORTER:-spec} --require tests/test-mocha.js tests/10-ManifestClient.spec.js",
-    "test-karma": "node tests/server1.js && node tests/server2.js && karma start karma.conf.js",
+    "test-servers": "node tests/servers.js",
+    "test-karma": "karma start karma.conf.js",
     "test-watch": "cross-env NODE_ENV=test mocha -r esm --watch --preserve-symlinks -t 30000 -A -R ${REPORTER:-spec} --require tests/test-mocha.js tests/*.spec.js",
     "coverage": "cross-env NODE_ENV=test nyc --reporter=lcov --reporter=text-summary npm run test-node",
     "coverage-ci": "cross-env NODE_ENV=test nyc --reporter=lcovonly npm run test-node",
